feature: rename Image base class to Asset
