--- src/fmgraphics/__init__.py
@@ -1,4 +1,0 @@
-from fmgraphics.image import Image
-
-
-__all__ = ["Image"]
